order_custom/serializers.py
from rest_framework import serializers

from order_custom.models import Purchase
from order_custom.models import Payment


class PurchaseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Purchase
        fields = '__all__'


class PaymentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Payment
        fields = '__all__'