requirements.txt
django-oscar
sorl-thumbnail
djangorestframework
markdown
django-filter

